added context menu for diseases cell in known drugs table, removed context menu for diseases table of pharm gkb
