+++ client/client/app/components/ngbTargetPanel/ngbDiseaseContextMenu/ngbDiseaseContextMenu.factory.js
@@ -1,0 +1,7 @@
+export default function(ngbContextMenuBuilder) {
+    return ngbContextMenuBuilder({
+        controller: 'ngbDiseaseContextMenuController',
+        controllerAs: '$ctrl',
+        template: require('./ngbDiseaseContextMenu.tpl.html')
+    });
+}
