+++ client/client/app/components/ngbTargetPanel/ngbDiseaseContextMenu/ngbDiseaseContextMenu.tpl.html
@@ -1,4 +1,4 @@
-<div class="ngb-diseases-table-context-menu">
+<div class="ngb-disease-context-menu">
     <md-list class="md-dense">
         <md-list-item>
             <md-button class="view-disease-details-button"
