--- client/client/app/components/ngbTargetPanel/ngbIdentificationsTab/ngbDiseasesPanel/ngbDiseasesTable/ngbDiseasesTableContextMenu/ngbDiseasesTableContextMenu.factory.js
@@ -1,7 +1,0 @@
-export default function(ngbContextMenuBuilder) {
-    return ngbContextMenuBuilder({
-        controller: 'ngbDiseasesTableContextMenuController',
-        controllerAs: '$ctrl',
-        template: require('./ngbDiseasesTableContextMenu.tpl.html')
-    });
-}
