+++ client/client/app/components/ngbTargetPanel/ngbIdentificationsTab/ngbKnownDrugsPanel/ngbDrugsTable/ngbDrugsTable.scss
@@ -97,6 +97,10 @@ ngb-drugs-table {
             color: inherit;
             text-decoration: none;
         }
+
+        &.disease {
+            width: 100%;
+        }
     }
 
     ngb-panel-error-list {
